feat(GroupPage): Add the Handler and the WebPage
GroupPageHandler.go will firstly check if you send him a 'post' request, if not it will redirect you to the homePage.
Next it will check if you send the 'id' in the right format, if it's not empty and if the 'id' exist. If not it will display an error.

GroupPage.gohtml : will display all the information of the given group