fix(Optimization): homePage was loading very slow due to coordinates request
Had to change the way the app requested the coordonates for each concerts, because when loading the homePage, it was getting all coordinates for each concerts of each groups present on the page.

Now, I maked a separated function 'getConcertsCoordinates' that do the same but only when needed (called).
So now, it's only called when loading a specific group page and not the homePage