feat(Filters Backend): Creation_Date range filter BackEnd
Filters.go :
I create the .go file 'Filters.go' that will have all the filters functions in it.
For now I have make the start of the 'FilterGroupsByCreationDate' filter, it take the 'creation date' sent, and compare it with all the groups present in the API, if there is one it put it in a filtered list of group and return it back.

HomePageHandler.go :
Added a 'if' condition that check if there is an actual 'POST' request, if there is it will start checking for each filters and after applying them, send the filtered list to the webServer.