feat(Filters Backend): Add of QtyOfMembers filter
HomePage.gohtml :
Add the QtyOfMembers form

Filters.go :
Add the filter, make a list that contain only groups who have the selected minimum of members and the maximum selected.

HomePageHandler.go :
Maked few change in the logic, the 'if' condition that was checking if the filter is not empty was not working because range form never send empty data.
After that I adapted the code so it can handle the 2 actual filter, when the first is done, the second one will filter the already filtered data.