fix(Browsers Request): Prevent the '/favicon.ico' request
Firefox and Chrome send 2 GET request while loading the website, 1 for the webpage itself and 1 to get the favicon.
But before my fix when a webbrowser was asking for the favicon, the webapp was loading ALL the code twice.

Now when it detect the path "/favicon.ico" it return directly