feat(SearchBar): Starting the predictive search (group name)
HomePage.gohtml :
- Linked the JavaScript
- Added an id to the search-bar to scope it in the JavaScript

HomePageHandler.go :
Now the handler will json marshall the data so it can be passed to the JavaScript

Structures.go :
- Changed/Added the jsons tag to make it more logic and to make it work with the JavaScript

SearchBar.js :
The future role of the script is to make suggestions depending of the user entry. The script will update the suggestion at each input.

For now it just check the input and search if there is correspondence  in the group's name