feat(SearchBar): Make the prediction work for Firefox
Before the change the prediction wasn't working on Firefox but only on Chrome.
It was because the 'dynamic' datalist markup don't work on Firefox.

To fix that i had the change the functionning of the 'makeSuggestion' for it to make UL - LI markups instead of the datalist

SearchResults.gohtml / HomePage.gohtml :
- Change the id's to be more conventional
- Delete the datalist

HomePage.css :
- Added all the CSS needed to have a stylized UL -LI

HomePageHandler.go :
- Delete the json function (it wasn't useful)

SearchBar.js :
- Recommented all the code to make it more readable
- Make the 'makeSuggestion' function work with Firefox by make UL - LI instead of datalist